+++ Sources/RepoBar/App/RepoBarApp.swift
@@ -89,6 +89,11 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
     func applicationShouldTerminateAfterLastWindowClosed(_: NSApplication) -> Bool {
         false
     }
+
+    func applicationShouldHandleReopen(_: NSApplication, hasVisibleWindows _: Bool) -> Bool {
+        self.menuManager?.openIssueNavigator()
+        return true
+    }
 }
 
 extension AppDelegate {