+++ Sources/RepoBar/StatusBar/MenuNotifications.swift
@@ -6,4 +6,8 @@ extension Notification.Name {
     static let menuDiagnosticsDidChange = Notification.Name("menuDiagnosticsDidChange")
     static let recentListFiltersDidChange = Notification.Name("recentListFiltersDidChange")
     static let gitHubReferenceMatchDidChange = Notification.Name("gitHubReferenceMatchDidChange")
+    static let issueNavigatorUseClipboard = Notification.Name("issueNavigatorUseClipboard")
+    static let issueNavigatorRefresh = Notification.Name("issueNavigatorRefresh")
+    static let issueNavigatorCopy = Notification.Name("issueNavigatorCopy")
+    static let issueNavigatorOpen = Notification.Name("issueNavigatorOpen")
 }